Remove test_utilities.go
